fix: support tree-shakable
