+++ lib/mixins/bindings.js
@@ -1,4 +1,4 @@
-const { getMapValueByKey } = require('../utils');
+const { getMapValueByKey } = require('../models-utils');
 
 /**
  * Implements functions to deal with the common Bindings object.
