+++ lib/mixins/external-docs.js
@@ -1,4 +1,4 @@
-const { getMapValueOfType } = require('../utils');
+const { getMapValueOfType } = require('../models-utils');
 
 const ExternalDocs = require('../models/external-docs');
