+++ lib/models/channel.js
@@ -1,4 +1,4 @@
-const { createMapOfType, getMapValueOfType, mix } = require('../utils');
+const { createMapOfType, getMapValueOfType, mix } = require('../models-utils');
 
 const Base = require('./base');
 const ChannelParameter = require('./channel-parameter');
