+++ lib/models/contact.js
@@ -1,5 +1,5 @@
   
-const { mix } = require('../utils');
+const { mix } = require('../models-utils');
 
 const Base = require('./base');
