+++ lib/utils.js
@@ -76,13 +76,6 @@ const findLocationOf = (keys, ast, initialFormat) => {
   };
 };
 
-const getMapValue = (obj, key, Type) => {
-  if (typeof key !== 'string' || !obj) return null;
-  const v = obj[String(key)];
-  if (v === undefined) return null;
-  return Type ? new Type(v) : v;
-};
-
 utils.tilde = (str) => {
   return str.replace(/[~\/]{1}/g, (m) => {
     switch (m) {
@@ -162,53 +155,6 @@ utils.toJS = (asyncapiYAMLorJSON) => {
       });
     }
   }
-};
-
-utils.createMapOfType = (obj, Type) => {
-  const result = {};
-  if (!obj) return result;
-
-  Object.entries(obj).forEach(([key, value]) => {
-    result[String(key)] = new Type(value);
-  });
-
-  return result;
-};
-
-utils.getMapValueOfType = (obj, key, Type) => {
-  return getMapValue(obj, key, Type);
-};
-
-utils.getMapValueByKey = (obj, key) => {
-  return getMapValue(obj, key);
-};
-
-/**
- * Extends a given model with additional methods related to frequently recurring models
- * @function mix
- * @private
- * @param  {Object} model model to extend
- * @param  {Array<Object>} mixins array with mixins to extend the model with
- */
-utils.mix = (model, ...mixins) => {
-  let duplicatedMethods = false;
-  function checkDuplication(mixin) {
-    // check duplication of model in mixins array
-    if (model === mixin) return true;
-    // check duplication of model's methods
-    duplicatedMethods = Object.keys(mixin).some(mixinMethod => model.prototype.hasOwnProperty(mixinMethod));
-    return duplicatedMethods;
-  }
-  
-  if (mixins.some(checkDuplication)) {
-    if (duplicatedMethods) {
-      throw new Error(`invalid mix function: model ${model.name} has at least one method that it is trying to replace by mixin`);
-    } else {
-      throw new Error(`invalid mix function: cannot use the model ${model.name} as a mixin`);
-    }
-  }
-  mixins.forEach(mixin => Object.assign(model.prototype, mixin));
-  return model;
 };
 
 utils.findRefs = (errors, initialFormat, asyncapiYAMLorJSON) => {
