+++ test/browser_test.js
@@ -61,4 +61,3 @@ describe('Check Parser in the browser', function() {
     }
   });
 });
-
